backend/pyproject.toml
[project]
name = "finance"
version = "0.1.0"
description = "Finance API"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.123.5",
    "psycopg2-binary>=2.9.11",
    "webauthn>=2.5.0",
    "PyJWT>=2.10.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.44",
    "uuid6>=2024.5.1",
    "uvicorn[standard]==0.32.0",
]

[dependency-groups]
dev = [
    "alembic>=1.17.2",
    "mypy>=1.19.1",
    "ruff>=0.14.8",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "A",     # シャドウ化されたビルトイン
    "ARG",   # 使われていない引数
    "ASYNC", # 非同期プログラミングのリンティング
    "B",     # 論理的なバグ
    "C4",    # 誤ったコンプリヘンション、辞書、リストなどの使用をキャッチ
    "C90",   # 関数の複雑度をチェック
    "COM",   # 末尾のカンマを強制（フォーマッタと衝突するものは無効化）
    "DTZ",   # datetimeで厳密なタイムゾーン操作を要求
    "E",     # PEP 8スタイルガイド (Error)
    "EXE",   # 実行可能ファイルに関するチェック
    "F",     # 明らかな論理エラーを検出 (Pyflakes)
    "FA",    # from __future__ import annotations を強制
    "FAST",  # FastAPI 特有のリンティング
    "FBT",   # 混乱を招くブール値引数
    "FLY",   # f-string への自動変換を推奨
    "FURB",  # モダンなPythonへのリファクタリング
    "G",     # ビルトインのロギングのより良い使用法
    "I",     # インポートのソート
    "INT",   # gettextの文字列フォーマットをチェック
    "ISC",   # 文字列連結の適切な使用
    "ICN",   # 一般的なインポート規約を使用
    "LOG",   # 標準ライブラリのロギングモジュール
    "N",     # 命名規則
    "PERF",  # パフォーマンス上の改善ポイント
    "PGH",   # 意図しないパターン
    "PIE",   # 一般的なコードの断片を簡略化
    "PL",    # Pylintの各種ルールを適用
    "PT",    # pytest 特有のリンティング
    "PTH",   # os.pathの代わりにpathlibを使用する
    "PYI",   # 型アノテーションのリンティング
    "Q",     # クォートのリンティング
    "RET",   # リターンプラクティス
    "RSE",   # raise文のスタイル
    "RUF",   # Ruff 独自のルール
    "S",     # セキュリティ診断
    "SIM",   # コードの簡略化
    "SLF",   # プライベート属性へのアクセス禁止
    "SLOT",  # メモリ節約のための __slots__ 使用
    "T10",   # デバッグ用の断点の使用を禁止
    "T20",   # print 文の使用を禁止
    "TID",   # インポートの整理・クリーンアップ
    "UP",    # 古いPython構文
    "W",     # PEP 8スタイルガイド (Warning)
    "YTT",   # Python のバージョン比較などの誤用
]

[tool.ruff.lint.per-file-ignores]
"scripts/**" = ["T201"]

[tool.mypy]
strict = true
explicit_package_bases = true
warn_unreachable = true

[tool.pyright]
venvPath = "."
venv = ".venv"
